requirements.txt
requests
python-dotenv
facebook-business